Rename evaluate-autocomplete to cody-bench
