+++ agent/src/cli/cody-bench/queries/cpp/context.scm
