+++ agent/src/cli/cody-bench/triggerAutocomplete.ts
@@ -11,7 +11,7 @@ import type { WrappedParser } from '../../../../vscode/src/tree-sitter/parser'
 import type { AutocompleteMatchKind } from './AutocompleteMatcher'
 import type { EvaluationDocument } from './EvaluationDocument'
 import type { TestParameters } from './TestParameters'
-import type { EvaluateAutocompleteOptions } from './evaluate-autocomplete'
+import type { EvaluateAutocompleteOptions } from './cody-bench'
 import { testParses } from './testParse'
 import { testTypecheck } from './testTypecheck'
